+++ wilmer-vega/src/main/java/gestiondeunbanco/wilmervega/domain/services/DeleteCompanyClient.java
@@ -1,5 +1,6 @@
 package gestiondeunbanco.wilmervega.domain.services;
 
+import gestiondeunbanco.wilmervega.domain.exceptions.NotFoundException;
 import gestiondeunbanco.wilmervega.domain.ports.CompanyClientPort;
 
 public class DeleteCompanyClient {
@@ -11,6 +12,9 @@ public class DeleteCompanyClient {
     }
 
     public void deleteById(Long id) {
+        if (companyClientPort.findById(id).isEmpty()) {
+            throw new NotFoundException("Cannot delete: CompanyClient not found with ID " + id);
+        }
         companyClientPort.deleteById(id);
     }
 }
