+++ wilmer-vega/src/main/java/gestiondeunbanco/wilmervega/domain/services/DeleteLoan.java
@@ -1,5 +1,6 @@
 package gestiondeunbanco.wilmervega.domain.services;
 
+import gestiondeunbanco.wilmervega.domain.exceptions.NotFoundException;
 import gestiondeunbanco.wilmervega.domain.ports.LoanPort;
 
 public class DeleteLoan {
@@ -11,6 +12,9 @@ public class DeleteLoan {
     }
 
     public void deleteById(Long id) {
+        if (loanPort.findById(id).isEmpty()) {
+            throw new NotFoundException("Cannot delete: Loan not found with ID " + id);
+        }
         loanPort.deleteById(id);
     }
 }
