+++ wilmer-vega/src/main/java/gestiondeunbanco/wilmervega/domain/services/DeleteNaturalClient.java
@@ -1,5 +1,6 @@
 package gestiondeunbanco.wilmervega.domain.services;
 
+import gestiondeunbanco.wilmervega.domain.exceptions.NotFoundException;
 import gestiondeunbanco.wilmervega.domain.ports.NaturalClientPort;
 
 public class DeleteNaturalClient {
@@ -11,6 +12,9 @@ public class DeleteNaturalClient {
     }
 
     public void deleteById(Long id) {
+        if (naturalClientPort.findById(id).isEmpty()) {
+            throw new NotFoundException("Cannot delete: NaturalClient not found with ID " + id);
+        }
         naturalClientPort.deleteById(id);
     }
 }
