+++ wilmer-vega/src/main/java/gestiondeunbanco/wilmervega/domain/services/DeleteTransfer.java
@@ -1,5 +1,6 @@
 package gestiondeunbanco.wilmervega.domain.services;
 
+import gestiondeunbanco.wilmervega.domain.exceptions.NotFoundException;
 import gestiondeunbanco.wilmervega.domain.ports.TransferPort;
 
 public class DeleteTransfer {
@@ -11,6 +12,9 @@ public class DeleteTransfer {
     }
 
     public void deleteById(Long id) {
+        if (transferPort.findById(id).isEmpty()) {
+            throw new NotFoundException("Cannot delete: Transfer not found with ID " + id);
+        }
         transferPort.deleteById(id);
     }
 }
