+++ wilmer-vega/src/main/java/gestiondeunbanco/wilmervega/domain/services/DeleteUser.java
@@ -1,5 +1,6 @@
 package gestiondeunbanco.wilmervega.domain.services;
 
+import gestiondeunbanco.wilmervega.domain.exceptions.NotFoundException;
 import gestiondeunbanco.wilmervega.domain.ports.UserPort;
 
 public class DeleteUser {
@@ -11,6 +12,9 @@ public class DeleteUser {
     }
 
     public void deleteById(Long id) {
+        if (userPort.findById(id).isEmpty()) {
+            throw new NotFoundException("Cannot delete: User not found with ID " + id);
+        }
         userPort.deleteById(id);
     }
 }
